Split StudyDeck container and presentation components
